- Remove QPlastiqueStyle usage
When compiling qjackctl from git with Qt 5.9 on Windows, there is an
error about QPlastiqueStyle: No such file or directory.

The offending code is:
https://github.com/rncbc/qjackctl/blob/aa73c02cdb81073c36efeb84805567676a612ddf/src/qjackctlMainForm.cpp#L63-L65
 #if defined(__WIN32__) || defined(_WIN32) || defined(WIN32)
 #include <QPlastiqueStyle>
 #endif

Where QPlastiqueStyle is used only here:
https://github.com/rncbc/qjackctl/blob/aa73c02cdb81073c36efeb84805567676a612ddf/src/qjackctlMainForm.cpp#L409-L411
 #if defined(__WIN32__) || defined(_WIN32) || defined(WIN32)
     QApplication::setStyle(new QPlastiqueStyle());
 #endif

Qt 5 seems to have replaced Plastique style with Fusion, and so
QPlastiqueStyle does not exists. This include is only made on Windows
given the ifdefs.

I think that include (and related usage of QPlastiqueStyle) should be
removed as Windows native style is working and will probably
look better integrated than a different custom style.