- Make Graph window behavior like other forms
Previously, the Graph window was always behaving like "Keep child windows
always on top" option was enabled even if it wasn't.

That is, the Graph window was always over the qjackctl main gui and on
Windows, there was no way to differentiate the main gui to the Graph
window in the taskbar.

Fix this by using the same constructor arguments as other forms.