[cmake] Use FindJack and FindPortAudio scripts to find dependencies even without pkg-config
This allows building qjackctl on Windows where pkg-config is not available
nor jack's pc file.

[cmake] Remove reference to qjackctl.rc and qjackctl.icns which don't exist

This issue was causing a build error on Windows with MinGW.
I've not tested on APPLE, but as images/qjackctl.icns does not exist
either, I expect it to fail too.

[cmake] install qjackctl.exe in bin dir

Previously, the executable file was not installed in the install directory,
but in the build directory.

If it is wanted to have qjackctl.exe at the root of the build directory
after a build, one can set RUNTIME_OUTPUT_DIRECTORY instead.

[cmake] Fix build on Windows with MSVC

_USE_MATH_DEFINES define is required to make M_PI available.
See https://docs.microsoft.com/fr-fr/cpp/c-runtime-library/math-constants?view=msvc-160