[cmake] enable package target
This will use zip format on Windows and tar.gz elsewhere.
The generated package for Windows should be redistributable as-is when
built with CONFIG_INSTALL_QT_DLLS enabled.

[windows] Rework installed files: only binaries and translations

Don't install manpages and spec file for Windows.
Install Qt's DLLs using windeployqt when CONFIG_INSTALL_QT_DLLS is ON.
Install and adjust translation path to match the one used by windeployqt.

windeployqt is a tool supplied with Qt which find which DLLs are required
for the application.