Fix missing connections with aliases (#196).
This fixes #196 by adjusting the call to findJackClientPort() in
qjackctlJackConnect::updateConnections(), so that it also considers the
aliases of the input ports if the target ports cannot be found by their
real name.

(This is done on the fly, using jack_port_get_aliases() on the spot. So
this will incur a tiny bit of additional overhead, but only if the
aliases display option is in effect.)