Fix the pretty name metadata display in the connections window.
We want this to be shown if it is non-empty, differs from the currently
shown alias, and the port item is not edited. To these ends, fix the
wrong test in qjackctlJackConnect.cpp:134 which needs to update the port
name (either the pretty name or the alias) if sPortNameEx != sPortName
*or* sPortNameEx != sPrettyName (not if both conditions are true at the
same time, as was done previously).