Fix the pretty name metadata display, part 3.
In our quest to improve alias/edit + metadata handling in the
connections and graph views, these fixes now hopefully ensure that the
connections view doesn't stomp on its own edits when the server is
restarted, so that aliases/edits take priority over pretty-name
metadata, as promised.

This should work now even if the server sets its own metadata at startup
(jack2, we're looking at you), and even if both the alias/edit and
metadata options are enabled in the display setup.